requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
python-dotenv>=1.0.0
langchain-ollama>=0.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
praw>=7.7.0
pytrends>=4.9.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
diskcache>=5.6.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=4.9.0
html5lib>=1.1
groq>=0.4.0
feedparser